AnalogInput Сделан рассчёт с коэффициентом

+++ Stm32/Drivers/AnalogInput.hpp
@@ -9,26 +9,26 @@
 class cAnalogInput : public iProcess
 {
 public:
-	cAnalogInput();
-	~cAnalogInput();
+  cAnalogInput();
+  ~cAnalogInput();
 
-	virtual void run();
+  virtual void run();
 
-	void Init();
-	void SetDataFromADC(uint32_t data);
-	uint32_t GetAverageData()const;
+  void Init();
+  void SetDataFromADC(uint32_t data);
+  uint32_t GetAverageData()const;
 
 private:
-	uint32_t m_Scale;
-	uint32_t m_DataRaw;
-	int32_t m_DataAcc;
-	uint8_t m_AvgDiv;
-	uint32_t m_DataAvg;
-	uint32_t m_DataScaling;
-	uint32_t m_MaxValue;
+  float m_Scale;
+  uint32_t m_DataRaw;
+  int32_t m_DataAcc;
+  uint8_t m_AvgDiv;
+  uint32_t m_DataAvg;
+  uint32_t m_DataScaling;
+  uint32_t m_MaxValue;
 
-	void CalcAverage(uint32_t data);
-	
+  void CalcAverage(uint32_t data);
+  
 };
 
 #endif /* DRIVERS_ANALOGINPUT_HPP_ */
